move wavelet decomp functions to wavelet_decomp module
